Isoler give-sample-name
